impl TODO list

+++ forum-management-system/src/main/java/com/forum/forummanagementsystem/services/UserServiceImpl.java
@@ -3,8 +3,10 @@ package com.forum.forummanagementsystem.services;
 import com.forum.forummanagementsystem.exceptions.AuthorizationException;
 import com.forum.forummanagementsystem.exceptions.EntityDuplicateException;
 import com.forum.forummanagementsystem.exceptions.EntityNotFoundException;
+import com.forum.forummanagementsystem.models.Admin;
 import com.forum.forummanagementsystem.models.FilterOptionsUser;
 import com.forum.forummanagementsystem.models.User;
+import com.forum.forummanagementsystem.repositories.interfaces.AdminRepository;
 import com.forum.forummanagementsystem.repositories.interfaces.UserRepository;
 import com.forum.forummanagementsystem.services.interfaces.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -16,17 +18,20 @@ import java.util.List;
 public class UserServiceImpl implements UserService {
 
     private static final String MODIFY_PROFILE_ERROR_MESSAGE = "Only profile owner can make changes to the profile.";
+    public static final String NON_ADMIN_ERROR = "Only admins can %s.";
 
     private final UserRepository userRepository;
+    private final AdminRepository adminRepository;
 
     @Autowired
-    public UserServiceImpl(UserRepository userRepository) {
+    public UserServiceImpl(UserRepository userRepository, AdminRepository adminRepository) {
+        this.adminRepository = adminRepository;
         this.userRepository = userRepository;
     }
 
     @Override
     public User getByUsername(String username) {
-       return userRepository.getByUsername(username);
+        return userRepository.getByUsername(username);
     }
 
     @Override
@@ -65,9 +70,37 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void deleteUser(User user) {
-        //TODO: check permissions
+    public void deleteUser(User user, User userAuthenticated) {
+        if (!userAuthenticated.isAdmin()) {
+            throw new AuthorizationException(String.format(NON_ADMIN_ERROR, "delete users"));
+        }
         userRepository.deleteUser(user);
+    }
+
+    @Override
+    public void blockUser(User userToBlock, User userAuthenticated) {
+        if (!userAuthenticated.isAdmin()) {
+            throw new AuthorizationException(String.format(NON_ADMIN_ERROR, "block users"));
+        }
+        userRepository.blockUser(userToBlock);
+    }
+
+    @Override
+    public void unblockUser(User userToUnblock, User userAuthenticated) {
+        if (!userAuthenticated.isAdmin()) {
+            throw new AuthorizationException(String.format(NON_ADMIN_ERROR, "unblock users"));
+        }
+        userRepository.unblockUser(userToUnblock);
+    }
+
+    @Override
+    public void makeAdmin(User userToMakeAdmin, User userAuthenticated) {
+        if (!userAuthenticated.isAdmin()) {
+            throw new AuthorizationException(String.format(NON_ADMIN_ERROR, "make user an admin"));
+        }
+        userToMakeAdmin.setAdmin(true);
+        userRepository.updateProfile(userToMakeAdmin);
+        adminRepository.makeAdmin(userToMakeAdmin);
     }
 
     private void checkModifyPermissions(User userAuthenticated, User userMapped) {
